refactor(add-bank-acc-form-context): rename AddBankAccFormContext file and fix form width
