+++ src/components/forms/UserInfoForm.tsx
@@ -2,7 +2,7 @@ import { CitySelect } from "@/components/inputs/CitySelect";
 import { DatePicker } from "@/components/inputs/DatePicker";
 import { DistrictSelect } from "@/components/inputs/DistrictSelect";
 import { Typography } from "@/components/ui/Typography";
-import { useAddBankAccFormContext } from "@/contexts/AddBankFormContext";
+import { useAddBankAccFormContext } from "@/contexts/AddBankAccFormContext";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { Box, Stack, TextField } from "@mui/material";
 import { useForm } from "react-hook-form";
