+++ src/pages/users/me/payments/new.tsx
@@ -1,7 +1,7 @@
 import { PageTitle } from "@/components/PageTitle";
 import { AddBankAccForm } from "@/components/forms/AddBankAccForm";
 import { BaseLayout } from "@/components/layouts/BaseLayout";
-import { AddBankAccFormContextProvider } from "@/contexts/AddBankFormContext";
+import { AddBankAccFormContextProvider } from "@/contexts/AddBankAccFormContext";
 import { Container } from "@mui/material";
 import Head from "next/head";
